add syntax parser
